Alex animation (#4)
* Updated Animal Movement to use the _Pelvis pivot

Needed for the fox to understand when it was back up to an edge. The a prefab with the Pelvis pivot attached should end up in the FoxAssets folder.

* Continuation of the edits to AnimalMovement.cs

If I remember correctly, this is where the AnimalMovement.cs script talks to the Fox's animator normally. Fairly certain this is necessary, though I'm not sure if I can explain how.

* Continuation of AnimalMovement.cs changes

This just adds the hash for the fallback condition in the fox's animator.

* Here's the Fox Baby Prefab

Not sure if I needed to add the .meta file, but oh well. If it's unneccessary, it hurts no one. If harmful, it can be deleted. And if it helps, better to have it, then.

* Slightly updated baby fox .prefab

Mostly I wanted to test how github reacts to two files of the same name being put in the same directory. For science.

* Added Updated Animal Fox.controller

Since I know there've been changes to this one, I want to see what github does when I upload this.

* Created the CustomAnims folder

Mostly just to see how to create folders within the online git client.

* Deleting CustomAnims.txt

Because that isn't what I wanted to do.

* Creating the CustomAnims folder

Properly this time.

* Adding the custom animations, 1 of 2

Everything from P-1 to P7.

* Adding in the custom animations, 2 of 2

It does seem that .meta files are important, so I'm glad I've been keeping them in.

* Updated the fox's animator...

... to disable looking at the player in all animations except P2 and P3.

* Delete gitkeep

I found it mildly humorous to keep it around about half a day. But it's not doing anything, so might as well get rid of it.

* Fixes the 45 degree offset the fox's pawprints had

* Reduces visibility of pawprints and step dust

So it looks less like a mini-Bikini Atoll nuclear test and more like a slightly dusty room. Also made the pawprints more transparent so it doesn't look like the fox is tracking evaporating mud all over the place.

* Because I'm not manually resolving every conflict

There's probably a better way to do this... but this'll probably work too.

* Changing the branch animator files.

For real this time.

* Amends Fox Controller to bring it up to date

This adds my changes on top of the existing changes from Build 3.

* Updates Fox Baby prefab under FoxAssets

It now has the AiAgent and the PelvisPivot + dust and track changes!

* Single change to fox controller

Forgot to make sure that P1 had an exit time. So this fixes that.

* Updated the Living Room scene to carry over fox changes

Doing it this way makes things simpler, I think.

* I previously forgot a transition

It was the exit transition for the fox's going-backwards-ledge catch-and-recovery animation. Silly me...