The RC Car is now a car, with wheel colliders and everything. It only occasionally gets its wheels trapped inside the colliders of other objects. It has a 4 wheel drive, steers with all 4 wheels, and the steering wheel is synced to the orientation of the controller.
Additionally, the changes to the Ambient and Fox audio scripts are in. The Ambient Audio now populates its lists off the contents of the requisite folders. The fox audio still has to be manually set up on the lists in a new gameobject called something like "FoxMasterAudioList" that has a bunch gameobjects below it that have the audio clips referenced by a script that neither initializes nor updates. All it does is point to the audio clips.

I also fixed a few collision boxes that the RC Car was getting stuck on. In general the RC Car should be functional, but it will occasionally be a tad... tempermental.