Made the fox's sound script automatically pull the lists from under the master list (apparently the order of the gameobjects listed in the heirarchy changes the order in which they get added to the sound script's internal list).
Also, added a very small script and audiosource to all the interactible objects so at least there's something there when they're thrown. The script only allows for a single sound and volume control. This does mean that when an object strikes a soft surface (the couch, the fox) that it sounds like its hit a hard surface. Hmmm. I'll add a driving sound to the RC car and then circle back around on that.