Alex bugfix (#11)
* This commit encompasses the bugfixes to the fox's prints, to the car, and to the ceiling. The lightmap was edited to add a single light in the corner, in addition.

* This includes the sounds of the fox (on grab and 'Play?'), and the ceiling fan.

* So this is the commit that includes all recommended changes for the pull request.
1#: sEText has been renamed to setText.
2#: Update method has been removed from FoxSounds because it wasn't being used.
3#: The generalized test object and its folder have been removed because of non-use.
4#: AmbientSound.cs now

* This includes changing a few variables to constants in each script, removing superflous comments in the scripts, and re-assigning the proper sounds to the three animations that have them.

* Added the comments to FoxSounds.cs to apologize and explain its (semi)-functionality.

* The RC Car is now a car, with wheel colliders and everything. It only occasionally gets its wheels trapped inside the colliders of other objects. It has a 4 wheel drive, steers with all 4 wheels, and the steering wheel is synced to the orientation of the controller.

Additionally, the changes to the Ambient and Fox audio scripts are in. The Ambient Audio now populates its lists off the contents of the requisite folders. The fox audio still has to be manually set up on the lists in a new gameobject called something like "FoxMasterAudioList" that has a bunch gameobjects below it that have the audio clips referenced by a script that neither initializes nor updates. All it does is point to the audio clips.

I also fixed a few collision boxes that the RC Car was getting stuck on. In general the RC Car should be functional, but it will occasionally be a tad... tempermental.

* Changed the car's collider so that the wheels don't get stuck inside of objects. Also, the collisions physics is currently continuous (dynamic). It can probably be reduced in accuracy to merely continuous with some testing.

* Made the fox's sound script automatically pull the lists from under the master list (apparently the order of the gameobjects listed in the heirarchy changes the order in which they get added to the sound script's internal list).

Also, added a very small script and audiosource to all the interactible objects so at least there's something there when they're thrown. The script only allows for a single sound and volume control. This does mean that when an object strikes a soft surface (the couch, the fox) that it sounds like its hit a hard surface. Hmmm. I'll add a driving sound to the RC car and then circle back around on that.

* So, all the interatible objects now have noises. The rc car goes vroom. And I added the rc car accelerate linearly over a variable-specified period of time.

The precise period of time might will probably need to be changed, but that's probably something someone else should test. My tolerances for VR-induced nausea are not representative of the target audiences'.

* Okay, this includes the things me and Ali talked about regarding the sound code. The requested changes are in, and I also slightly changed the lightmap. All audio and light gameobjects are under empty game objects at the bottom that are respectively titled.

* Whoops. Forgot the null check. So that's now in the AmbientSound script.

* Added the prefabs for the MasterAudioGroup and MasterLightGroup (in space nobody can hear you die of embarrassment.)