Basically, I touched a lot of things related to the AnimalAIControl and ActionZone scripts. A lot of things. I also put the Fox AI Agent further forward on the fox.
The design principle I'm adopting for making the Fox change it's interactions with items like the jack in the box is sort of in a sort of a lego style. Creating events that tell the fox to go to other events so that an illusion of intentionality (on the fox's part is achieved).

The actual details are that I've given ActionZones additional events based on the fox's first seeing the object in question, on enabling the event, and on grabbing the object. I've also added a function to copy these events + the normal ones for OnEnter, OnAction, and OnExit from dummy gameobjects in order to give a single Actionzone multiple "states".
I've also added a override to the AnimalAIControl script, where, once triggered, it ignores anything that doesn't also pass along a boolean value of "true".

Ultimately, this is the initial rough version that I've made work in a "at all costs" sort of way. I intend to do a performance and cleanup run on the code to take care of my messes. I just want this version on Git in case I change something, everything breaks, and I can't tell why.