So, TriggerJawOverride and TriggerHeadOverride have been added to the AnimalAIControl script.
Each respectively triggers one of the two new layers on the Fox's animator. Since we only have a head shaking and a jaw open animation for right now, that the only animation each triggers.

TriggerJawOverride sets the jaw to a float between 0 and 1, where 1 is the fox's mouth staying open staying open completely with no influence from other animations while 0 is the opposite. It smoothly transitions between the new value and the previous value over .5 seconds.

TriggerHeadOverride is slightly different. It requires a fade in, stay, and fade out time. Likewise, this function also smoothly transitions.