Okay, this should be the final changes to animal pathing and whatnot before we (attempt) to merge all of this into the master branch. I've removed all instances of using Agent.calculatePath() since it wasn't giving reliable results anyway.
PathingTimeOut() has been rewritten to take this into account. It now checks for 2.5 seconds to see if the fox can find a path to a target. If it can't, it times out. If it finds a path, but loses it within the next 2.5 seconds, then it tries to find the path for 2.5 seconds again. If this cycle goes over around 12 seconds or so, it times out. On timing out, it always interrupts the fox's pathing so it can shake its head, calls SetTarget(null), and calls VoiceFox(5).

I also made the area underneath the Table with the two chairs pathable. And moved around the toys so they don't all make falling sounds when the game starts. Then there's Ben's changes to that the fox properly opens its mouth for the toys and whatnot. So that's all fine.

I think that's everything...