Included a state boolean to keep Agent.destination from being set once per frame. Lowered the interrupt pathing interval when the fox can't find a path so it looks a little bit more natural.
Lowered the position of a few event transforms so that the fox properly registers being able to path to them.

Included a timer delay in Foxsounds so that it can only trigger so many times per second. Not strictly necessary, but it's more fail friendly.

Properly ensured RelativeHeadRotate script could change targets.