Okay, this is the final commit for this version. Everything has been updated to play well with each other. Everything seems to work and performance does not seem to be really be impacted on my Samsung Galaxy S7.
List of changes in this include:

Made it so that switching channels with the remote is a simple matter of swiping left or right, rather than rotating around the center.
Made ClockHours only checks system time once per hour. Previously it called it once per second, which seemed to have a performance impact.
Updated the check in TabletVisibility to actually work as intended. (Previously, returned true if the player's CURRENTSTATE wasn't Driving OR it wasn't Remote. i.e. it always returned true.
Made lookTarget private in RelativeHeadRotate, as it should be (it's only referenced internally, and it will only be referenced internally).
Removed superfluous variables in FoxItemInteraction && made it so that the fox doesn't collide with whatever it drops for 1.5 seconds.
Updated the ActionZone script so that the fox wouldn't hit (& knock away) objects that it's currently animating next to.
Updated FoxTargetManager so that, if an object was previously thrown, the fox wouldn't exit the ReadyToPlay animation immediately.
Added a new layer that only collides with animals, and used it to set up some blockers around the table so the fox wouldn't accidentally get stuck behind it.
Made raycasts not hit triggers. I'm pretty sure that used to be disabled. Well, now it's disabled again.
Gave the ReadyToPlay(In) animation several different transitions out with different exit times. Anything else just looked wonky.

Of course, everything else to date is also included. It's a grab bag. It's a variety pack.