Tidied up ActionZone by combining a number of if statements. Also removed the equals and gethashcode override.
Fixed an issue with AnimalAIControl where it would enable and disable the AIAgent rapidly (no idea how it happened, but it did)
Also moved the evaluation of the foxes path into the PathingEvaluate function which is called at the end of UpdateAgent(). Added an enum, enum variable, and two floats for that purpose.