The minor changes and bugfixes.
The ball has been given a higher poly mesh and the material has been slightly darkened to reduce the obvious white circle caused by the interaction between VR and MSAA. Apparently this is a known issue that would require a large amount of computation to solve in a non-hacky manney. It exists everywhere, but is most noticeable on the ball due to the fact that you're almost always looking near-parallel to its surface normal.

The table has also had a box added under it, so that it is clearly communicated that it is outside the limits of play.

An attributions rtf has been added to the _VrPetAssets folder for when we have to make a credits page.

The steering script has been edited so that movement forward and backward is controlled by pad of the controller, rather than the trigger. Instead, the trigger instead plays a car horn.

A tableticon and button for the jack has been added.

The Couchbutton script has been updated with a null check for the offmeshlinks it disables.

The RcDriverButtonClick script has been updated with a reference to the StandardLaserPointer attached to the pointer, in order to disable it while in the car.

A channelpopulator script has been added to the TV. It takes twelve channel names from a specified static url and uses them to populate the TV's channel button menu.