A few more quick fixes.
The fox no longer tries to look at things while it sleeps.

Changed a few things to make sure that the fox disables its target override when set to wander, and that firsttimetrigger is properly set to false every time an actionzone is enabled.

Also updated the tablet prefab with a reference to the laser properly.