Added the new tablet model and centralized the Tablet's functions in the TabletFunctions script, including control over the transitions to and from the couch and car. Additional functions include control over swinging the side panel of the Tablet out, and all functions regarding the control of the volume for the Ambient, Music, and Master audio channels.
Additionally, the UIMenuStateManager and UpdateSliderCount scripts have been created to assist in the TabletFunction scripts. The UIMenuStateManager script largely borrows functionality from the UIMenuStateManager, and is a relabeling of the same. The Television has been updated to the UIMenuStateManager script instead.

The ModelButtonMovement script has been created for the buttons of the Tablet. They are triggered off one of the EasyInputforGearVR scripts that were attached to each button.

As part of the audio controls available via the tablet, the AmbientSound script has been updated make that control possible. The AmbientSoundManager script has also been created to centralize that control.

The OVRScreenFade script has been edited to expose the isFading bool so that the TabletFunctions script does not trigger multiple, contradictory transitions simultaneously.

Rather than setting the ActionZone.enabled to true or false, AnimalAIControl and Actionzone instead set the value of the readyToTrigger bool. Setting enabled = true/false was interfering with the overall function of the script and leading to volatile behaviour.

A height check has been added to the Update function of AnimalAIControl, in case the fox falls through the floor or is somehow launched into the ceiling.