Added the new tablet icons from Ainchase and hooked them up appropriately.
Moved the TabletFunctions, UIMenuStateManager, UpdateSliderCount, ModelButtonMovement, ChannelPopulator, AmbientSoundManager scripts to the appropriate folder.

Added assets for in-game debug purposes. They are currently disabled, but present, in the scene. Currently debug actions include forcing the window and sleep events and pointing out where in the scene the fox is.

From the Actionzone script, the whole grapdropinvokingcycle schtick has been removed. It was ultimately pointless; a duplication of efforts. Also, when an actionzone is picked up, the script now also properly resets the firstTimeTrigger bool.

The following few lines are about the AnimalAIControl script.

The height check has been limited to checking whether the fox's y position is below -20, for better compatibility with hypothetical future scenes.

The way that the interrupt timer keeps the fox from moving has been changed. Rather than keep the fox from doing any pathing-related checks, it instead merely sets the Direction Vector3 to (0, 0, 0).

The evaluate pathing check has been altered so that there is now an absolute cap on moving to perform an action of 16 seconds. If the fox cannot manage to arrive in that time, it's probably safe to say that it ain't happening. The pathing check will also continue to happen even if a valid path has been found and maintained for 4 seconds.

The SlowWalk pathing state now sets the animation speed to exactly .7f at all times. .3f, while still perfectly serviceable, was too slow for general purposes.

The SetTarget function has been changed to always invoke the onEnable events of an attached ActionZone script, regardless of whether or not the target is set.